Rename batch modules
