+++ packages/slate/src/core/apply.ts
@@ -2,7 +2,10 @@ import { WithEditorFirstArg } from '../utils/types'
 import { Editor } from '../interfaces/editor'
 import { runOperation } from './apply-operation'
 import { isBatching, scheduleOnChange } from './batch'
-import { applyOperationBatch, applyOperationInBatch } from './batching/executor'
+import {
+  applyOperationBatch,
+  applyOperationInBatch,
+} from './batching/apply-operation-batch'
 
 // Keep the public apply seam boring: single-op execution lives here, while
 // batch planning and execution stay under ./batching/.
